fix VC6 does not define `intptr_t`; revert change and use `long`; maint!?

+++ lglob.h
@@ -69,7 +69,7 @@
                     char dir[_MAX_DIR]; \
                     char fname[_MAX_FNAME]; \
                     char ext[_MAX_EXT]; \
-                    intptr_t handle;
+                    long handle;
 
 #else
 #if MSDOS_COMPILER==WIN32C && !defined(_MSC_VER) /* Borland C for Windows */
